requirements.txt
streamlit
agno>=2.2.10
openai
icalendar
google-search-results
googlesearch-python
pycountry
mcp
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
